Mount `/app` with host user as owner and run Gradle with different cache directory
This commit mounts the `/app` volume with `appuser` as the owner, which
we assume to be the host machine's current user (uid=1000).

Crucially, we also need to run the Gradle wrapper with
`--project-cache-dir /tmp/gradle-cache` to avoid the project `.gradle`
directory being used. This avoids cache lock contention which would
otherwise stop us from building the project from inside IntelliJ or
working with the project outside of Docker.